LPD-46294 move cache to FriendlyURLSeparatorConfigurationManager
