+++ modules/apps/friendly-url/friendly-url-service/src/main/java/com/liferay/friendly/url/internal/provider/FriendlyURLSeparatorProviderImpl.java
@@ -7,17 +7,11 @@ package com.liferay.friendly.url.internal.provider;
 
 import com.liferay.friendly.url.configuration.manager.FriendlyURLSeparatorConfigurationManager;
 import com.liferay.friendly.url.provider.FriendlyURLSeparatorProvider;
-import com.liferay.portal.kernel.cache.MultiVMPool;
-import com.liferay.portal.kernel.cache.PortalCache;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 
-import java.util.Map;
-
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
 
 /**
@@ -30,17 +24,9 @@ public class FriendlyURLSeparatorProviderImpl
 	@Override
 	public String getFriendlyURLSeparator(long companyId, String key) {
 		try {
-			JSONObject jsonObject = _portalCache.get(companyId);
-
-			if (jsonObject != null) {
-				return jsonObject.getString(key);
-			}
-
 			JSONObject friendlyURLSeparatorsJSONObject =
 				_friendlyURLSeparatorConfigurationManager.
 					getFriendlyURLSeparatorsJSONObject(companyId);
-
-			_portalCache.put(companyId, friendlyURLSeparatorsJSONObject);
 
 			return friendlyURLSeparatorsJSONObject.getString(key);
 		}
@@ -53,29 +39,11 @@ public class FriendlyURLSeparatorProviderImpl
 		return null;
 	}
 
-	@Activate
-	protected void activate(Map<String, Object> properties) {
-		_portalCache =
-			(PortalCache<Long, JSONObject>)_multiVMPool.getPortalCache(
-				FriendlyURLSeparatorProvider.class.getName());
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		_multiVMPool.removePortalCache(
-			FriendlyURLSeparatorProvider.class.getName());
-	}
-
 	private static final Log _log = LogFactoryUtil.getLog(
 		FriendlyURLSeparatorProviderImpl.class.getName());
 
 	@Reference
 	private FriendlyURLSeparatorConfigurationManager
 		_friendlyURLSeparatorConfigurationManager;
-
-	@Reference
-	private MultiVMPool _multiVMPool;
-
-	private PortalCache<Long, JSONObject> _portalCache;
 
 }
